backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
ultralytics==8.0.206
PyMuPDF==1.23.8
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
